requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
distlib==0.3.2
filelock==3.0.12
grpcio==1.38.0
grpcio-testing==1.38.0
grpcio-tools==1.38.0
iniconfig==1.1.1
mock==4.0.3
packaging==20.9
pbr==5.6.0
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-asyncio==0.15.1
six==1.16.0
testresources==2.0.1
toml==0.10.2
virtualenv==20.4.7
